moved ptplayer from utils to managers because it's a manager
