+++ src/ace/managers/ptplayer.c
@@ -9,7 +9,7 @@
 // - http://coppershade.org/articles/More!/Topics/Protracker_Tempo
 // - http://coppershade.org/articles/More!/Topics/Protracker_File_Format/
 
-#include <ace/utils/ptplayer.h>
+#include <ace/managers/ptplayer.h>
 #include <ace/managers/log.h>
 #include <ace/managers/system.h>
 #include <ace/utils/custom.h>
